Fix agent crash on ARM64 (Raspberry Pi)
The function ILibWebClient_ProcessWebSocketData in
microstack/ILibWebClient.c reads a 64-bit integer from an address that
is only 2-byte aligned. ARM64 kernels support unaligned accesses from
userspace, but it is not enabled by default and frowned upon - the
required exception handling is ugly and inefficient.

Fix that illegal access in a simple way with a memcpy. More involved and
efficient solutions are also available, such as replacing the byte order
conversion functions with de/serialise functions that take a pointer.

Fixes: https://github.com/Ylianst/MeshAgent/issues/183

Signed-off-by: Phil Elwell <phil@raspberrypi.com>